fit_evotree

EvoTrees.fit_evotree — Function

fit_evotree(params, X_train, Y_train, W_train=nothing;
    X_eval=nothing, Y_eval=nothing, W_eval = nothing,
    early_stopping_rounds=9999,
    print_every_n=9999,
    verbosity=1)

Main training function. Performs model fitting given configuration params, X_train, Y_train input data.

Arguments

params::EvoTypes: configuration info providing hyper-paramters. EvoTypes comprises EvoTreeRegressor, EvoTreeClassifier, EvoTreeCount or EvoTreeGaussian

X_train::Matrix: training data of size [#observations, #features].

Y_train::Vector: vector of train targets of length #observations.

W_train::Vector: vector of train weights of length #observations. Defaults to nothing and a vector of ones is assumed.

Keyword arguments

X_eval::Matrix: evaluation data of size [#observations, #features].

Y_eval::Vector: vector of evaluation targets of length #observations.

W_eval::Vector: vector of evaluation weights of length #observations. Defaults to nothing (assumes a vector of 1s).

early_stopping_rounds::Integer: number of consecutive rounds without metric improvement after which fitting in stopped.

print_every_n: sets at which frequency logging info should be printed.

verbosity: set to 1 to print logging info during training.

Predict

MLJModelInterface.predict — Function

predict(loss::L, tree::Tree{T}, X::AbstractMatrix, K)

Prediction from a single tree - assign each observation to its final leaf.

predict(model::GBTree{T}, X::AbstractMatrix)

Predictions from an EvoTrees model - sums the predictions from all trees composing the model.

Features Importance

EvoTrees.importance — Function

importance(model::GBTree, vars::AbstractVector)

Sorted normalized feature importance based on loss function gain.