EvoTrees.jl supports four model families:

EvoTreeRegressor

Linear (minimize mean-squared error)

Logistic (minimize cross-entropy)

L1: minimize mean-absolute error

Quantile

EvoTreeClassifier

EvoTreeCount

EvoTreeGaussian

EvoTrees.EvoTreeRegressor — Type

EvoTreeRegressor(;
    T::Type=Float64,
    loss=:linear,
    nrounds=10,
    λ=0.0,
    γ=0.0,
    η=0.1,
    max_depth=5,
    min_weight=1.0,
    rowsample=1.0,
    colsample=1.0,
    nbins=64,
    α=0.5,
    metric=:mse,
    rng=123,
    device="cpu")

Docs for regressor.

EvoTrees.EvoTreeClassifier — Type

EvoTreeClassifier

A model type for constructing a evo tree classifier, based on EvoTrees.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

EvoTreeClassifier = @load EvoTreeClassifier pkg=EvoTrees

Do model = EvoTreeClassifier() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeClassifier(loss=...).

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:OrderedFactor or <:Multiclass; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

loss: One of :linear, :logistic, :quantile, :L1

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeClassifier = @load EvoTreeClassifier pkg=EvoTrees
model = EvoTreeClassifier(max_depth=5, num_bins=32)
X, y = @load_crab
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.

EvoTrees.EvoTreeCount — Type

EvoTreeCount()

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:OrderedFactor or <:Multiclass; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

loss: One of :linear, :logistic, :quantile, :L1

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeCount = @load EvoTreeCount pkg=EvoTrees
model = EvoTreeCount(max_depth=5, num_bins=32)
X, y = @load_crab
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.

EvoTrees.EvoTreeGaussian — Type

EvoTreeGaussian

A model type for constructing a evo tree gaussian, based on EvoTrees.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

EvoTreeGaussian = @load EvoTreeGaussian pkg=EvoTrees

Do model = EvoTreeGaussian() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeGaussian(loss=...).

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:OrderedFactor or <:Multiclass; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeGaussian = @load EvoTreeGaussian pkg=EvoTrees
model = EvoTreeGaussian(max_depth=5, num_bins=32)
X, y = @load_crab
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.