EvoTrees.jl supports four model families:

EvoTreeRegressor

Linear (minimize mean-squared error)

Logistic (minimize cross-entropy)

L1: minimize mean-absolute error

Quantile

EvoTreeClassifier

EvoTreeCount

EvoTreeGaussian

EvoTreeRegressor

EvoTrees.EvoTreeRegressor — Type

EvoTreeRegressor

A model type for constructing a EvoTree regressor, based on EvoTrees.jl, and implementing both an internal API the MLJ model interface.

Hyper-parameters

loss=:linear: One of :linear, :logistic, :quantile, :L1.

nrounds=10: Number of rounds. It corresponds to the number of trees that will be sequentially stacked.

lambda::T=0.0 L2 regularization term on weights. Must be >= 0. Higher lambda can result in a more robust model.

gamma::T=0.0 Minimum gain imprvement needed to perform a node split. Higher gamma can result in a more robust model.

alpha::T=0.5 Loss specific parameter in the [0-1] range: - :quantile: target quantile for the regression. - :L1: weighting parameters to positive vs negative residuals. - Positive residual weights = alpha - Negative residual weights = (1 - alpha)

max_depth=5 Maximum depth of a tree. Must be >= 1. A tree of depth 1 is made of a single prediction leaf. A complete tree of depth N contains 2^(depth - 1) terminal leaves and 2^(depth - 1) - 1 split nodes. Compute cost is proportional to 2^max_depth. Typical optimal values are in the [3-9] range.

min_weight=0.0 Minimum weight needed in a node to perform a split. Matches the number of observations by default or the sum of weights as provided by theweights` vector.

rowsample=1.0 Proportion of rows that are sampled at each iteration to build the tree. Should be ]0, 1].

colsample=1.0 Proprtion of columns / features that are sampled at each iteration to build the tree. Should be ]0, 1].

nbins=32 Number of bins into which each feature is quantized. Buckets are defined based on quantiles, hence resulting in equal weight bins.

rng=123: Either an integer used as a seed to the random number generator or an actual random number generator (::Random.AbstractRNG).

metric::Symbol=mse: Metric that is to be tracked during the training process. One of: :none, mse, mae, logloss.

device="cpu" Hardware device to use for computations. Can be either "cpu" or "gpu".

Internal API

Do params = EvoTreeRegressor() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeRegressor(loss=...).

Training model

A model is built using fit_evotree: fit_evotree(params, X_train, Y_train, W_train=nothing; kwargs...).

MLJ Interface

From MLJ, the type can be imported using:

EvoTreeRegressor = @load EvoTreeRegressor pkg=EvoTrees

Do model = EvoTreeRegressor() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeRegressor(loss=...).

Training model

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:Continuous; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

:fitresult: The GBTree object returned by EvoTrees.jl fitting algorithm.

Report

The fields of report(mach) are:

:feature_importances: Feature importances based on the gain brought at each node split in the form of a Vector{Pair{String, Float64}}.

Examples

using EvoTrees, DataFrames
params = EvoTreeRegressor(max_depth=5, nbins=32, nrounds=100)
nobs, nfeats = 1_000, 5
X, y = randn(nobs, nfeats), rand(nobs)
model = fit_evotree(params, X, y)
preds = EvoTrees.predict(model, X)

using MLJ
EvoTreeRegressor = @load EvoTreeRegressor pkg=EvoTrees
model = EvoTreeRegressor(max_depth=5, nbins=32, nrounds=100)
X, y = @load_boston
mach = machine(model, X, y) |> fit!
preds = predict(mach, X)

EvoTreeClassifier

EvoTrees.EvoTreeClassifier — Type

EvoTreeClassifier

A model type for constructing a evo tree classifier, based on EvoTrees.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

EvoTreeClassifier = @load EvoTreeClassifier pkg=EvoTrees

Do model = EvoTreeClassifier() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeClassifier(loss=...).

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:Finite; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeClassifier = @load EvoTreeClassifier pkg=EvoTrees
model = EvoTreeClassifier(max_depth=5, num_bins=32)
X, y = @load_iris
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.

EvoTreeCount

EvoTrees.EvoTreeCount — Type

EvoTreeCount

A model type for constructing a evo tree count, based on EvoTrees.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

EvoTreeCount = @load EvoTreeCount pkg=EvoTrees

Do model = EvoTreeCount() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeCount(loss=...).

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:Count; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeCount = @load EvoTreeCount pkg=EvoTrees
model = EvoTreeCount(max_depth=5, num_bins=32)
X, y = @load_crab
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.

EvoTreeGaussian

EvoTrees.EvoTreeGaussian — Type

EvoTreeGaussian

A model type for constructing a evo tree gaussian, based on EvoTrees.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

EvoTreeGaussian = @load EvoTreeGaussian pkg=EvoTrees

Do model = EvoTreeGaussian() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in EvoTreeGaussian(loss=...).

Training data

In MLJ or MLJBase, bind an instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns each have one of the following element scitypes: Continuous, Count, or <:OrderedFactor; check column scitypes with schema(X)

y: is the target, which can be any AbstractVector whose element scitype is <:Continuous; check the scitype with scitype(y)

Train the machine using fit!(mach, rows=...).

Hyper-parameters

nrounds=10: Max number of rounds

rng=Random.GLOBAL_RNG: random number generator or seed

Operations

predict(mach, Xnew): return predictions of the target given features Xnew having the same scitype as X above. Predictions are deterministic.

Fitted parameters

The fields of fitted_params(mach) are:

gbtree: The GBTree object returned by EvoTrees.jl fitting algorithm

Report

The fields of report(mach) are:

...

Examples

using MLJ
EvoTreeGaussian = @load EvoTreeGaussian pkg=EvoTrees
model = EvoTreeGaussian(max_depth=5, num_bins=32)
X, y = @load_crab
mach = machine(model, X, y) |> fit!

See also EvoTrees.jl.